+++ src/typecode/magic2.py
@@ -1,26 +1,13 @@
 #
-# Copyright (c) nexB Inc. and others.
-# SPDX-License-Identifier: Apache-2.0 AND MIT
-#
-# Visit https://aboutcode.org and https://github.com/nexB/ for support and download.
+# Copyright (c) nexB Inc. and others. All rights reserved.
 # ScanCode is a trademark of nexB Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
+# SPDX-License-Identifier: Apache-2.0 AND MIT
+# See http://www.apache.org/licenses/LICENSE-2.0 for the license text.
+# See https://github.com/nexB/typecode for support or download.
+# See https://aboutcode.org for more information about nexB OSS projects.
 #
 # This code was in part derived from the python-magic library:
-# The MIT License (MIT)
-#
+
 # Copyright (c) 2001-2014 Adam Hupp
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -42,20 +29,38 @@
 # SOFTWARE.
 
 import ctypes
+import glob
 import os
 
 from commoncode import command
-from plugincode.location_provider import get_location
-
-from os import fsencode
-
-TRACE = False
+from commoncode.system import on_windows
+import warnings
 
 """
 magic2 is minimal and specialized wrapper around a vendored libmagic file
 identification library. This is NOT thread-safe. It is based on python-magic
 by Adam Hup and adapted to the specific needs of ScanCode.
 """
+
+# Tracing flag
+TRACE = True
+
+
+def logger_debug(*args):
+    pass
+
+
+if TRACE:
+    import logging
+    import sys
+
+    logger = logging.getLogger(__name__)
+    logging.basicConfig(stream=sys.stdout)
+    logger.setLevel(logging.DEBUG)
+
+    def logger_debug(*args):
+        return logger.debug(' '.join(isinstance(a, str) and a or repr(a) for a in args))
+
 #
 # Cached detectors
 #
@@ -74,24 +79,184 @@ DETECT_MIME = MAGIC_NONE | MAGIC_MIME
 DETECT_ENC = MAGIC_NONE | MAGIC_MIME | MAGIC_MIME_ENCODING
 
 # keys for plugin-provided locations
-TYPECODE_LIBMAGIC_LIBDIR = 'typecode.libmagic.libdir'
 TYPECODE_LIBMAGIC_DLL = 'typecode.libmagic.dll'
-TYPECODE_LIBMAGIC_DATABASE = 'typecode.libmagic.db'
+TYPECODE_LIBMAGIC_DB = 'typecode.libmagic.db'
+
+TYPECODE_LIBMAGIC_PATH_ENVVAR = 'TYPECODE_LIBMAGIC_PATH'
+TYPECODE_LIBMAGIC_DB_PATH_ENVVAR = 'TYPECODE_LIBMAGIC_DB_PATH'
+
+
+class NoMagicLibError(Exception):
+    """
+    Raised when no libmagic library is found.
+    """
+
+
+def load_lib_failover():
+    """
+    Return a loaded libmagic from well known system installation locations.
+    This is a function originally from python-magic.
+    """
+    libmagic = None
+    # Let's try to find magic or magic1
+    dll = (
+        ctypes.util.find_library('magic')
+        or ctypes.util.find_library('magic1')
+        or ctypes.util.find_library('cygmagic-1')
+        or ctypes.util.find_library('libmagic-1')
+        # for MSYS2
+        or ctypes.util.find_library('msys-magic-1')
+    )
+    # necessary because find_library returns None if it doesn't find the library
+    if dll:
+        libmagic = ctypes.CDLL(dll)
+
+    if not (libmagic and libmagic._name):
+        windows_dlls = [
+            'magic1.dll',
+            'cygmagic-1.dll',
+            'libmagic-1.dll',
+            'msys-magic-1.dll',
+        ]
+        platform_to_lib = {
+            'darwin': (
+                [
+                    '/opt/local/lib/libmagic.dylib',
+                    '/usr/local/lib/libmagic.dylib',
+                ] +
+                # Assumes there will only be one version installed when using brew
+                glob.glob('/usr/local/Cellar/libmagic/*/lib/libmagic.dylib')
+            ),
+            'win32': windows_dlls,
+            'cygwin': windows_dlls,
+            'linux': ['libmagic.so.1'],
+        }
+        # fallback for some Linuxes (e.g. Alpine) where library search does not
+        # work # flake8:noqa
+        platform = 'linux' if sys.platform.startswith('linux') else sys.platform
+        for dll in platform_to_lib.get(platform, []):
+            try:
+                libmagic = ctypes.CDLL(dll)
+                break
+            except OSError:
+                pass
+
+    if libmagic and libmagic._name:
+        return libmagic
 
 
 def load_lib():
     """
-    Return the loaded libmagic shared library object from plugin-provided path.
+    Return the libmagic shared library object loaded from either:
+    - an environment variable ``TYPECODE_LIBMAGIC_PATH``
+    - a plugin-provided path,
+    - the system PATH.
+    Raise an NoMagicLibError if no libmagic can be found.
     """
-    dll = get_location(TYPECODE_LIBMAGIC_DLL)
-    libdir = get_location(TYPECODE_LIBMAGIC_LIBDIR)
-    if not (dll and libdir) or not os.path.isfile(dll) or not os.path.isdir(libdir):
-        raise Exception(
-            'CRITICAL: libmagic DLL and is magic database are not installed. '
+    from plugincode.location_provider import get_location
+
+    # try the environment first
+    dll_loc = os.environ.get(TYPECODE_LIBMAGIC_PATH_ENVVAR)
+
+    if TRACE and dll_loc:
+        logger_debug('load_lib:', 'got environ magic location:', dll_loc)
+
+    # try a plugin-provided path second
+    if not dll_loc:
+        dll_loc = get_location(TYPECODE_LIBMAGIC_DLL)
+
+        if TRACE and dll_loc:
+            logger_debug('load_lib:', 'got plugin magic location:', dll_loc)
+
+    # try well known locations
+    if not dll_loc:
+        failover_lib = load_lib_failover()
+        if failover_lib:
+            warnings.warn(
+                'System libmagic found in typical location is used. '
+                'Install instead a typecode-libmagic plugin for best support.'
+            )
+            return failover_lib
+
+    # try the PATH
+    if not dll_loc:
+        dll = 'libmagic.dll' if on_windows else 'libmagic.so'
+        dll_loc = command.find_in_path(dll)
+
+        if dll_loc:
+            warnings.warn(
+                'libmagic found in the PATH. '
+                'Install instead a typecode-libmagic plugin for best support.'
+            )
+
+        if TRACE and dll_loc:
+            logger_debug('load_lib:', 'got path magic location:', dll_loc)
+
+    if not dll_loc or not os.path.isfile(dll_loc):
+        raise NoMagicLibError(
+            'CRITICAL: libmagic DLL and its magic database are not installed. '
             'Unable to continue: you need to install a valid typecode-libmagic '
-            'plugin with a valid and proper libmagic and magic DB available.'
+            'plugin with a valid and proper libmagic and magic DB available.\n'
+            f'OR set the {TYPECODE_LIBMAGIC_PATH_ENVVAR} and '
+            f'{TYPECODE_LIBMAGIC_DB_PATH_ENVVAR} environment variables.\n'
+            f'OR install libmagic in typical common locations.\n'
+            f'OR have a libmagic in the system PATH.\n'
     )
-    return command.load_shared_library(dll, libdir)
+    return command.load_shared_library(dll_loc)
+
+
+def get_magicdb_location(_cache=[]):
+    """
+    Return the location of the magicdb loaded from either:
+    - an environment variable ``TYPECODE_LIBMAGIC_DB_PATH``,
+    - a plugin-provided path,
+    - the system PATH.
+    Trigger a warning if no magicdb file is found.
+    """
+    if _cache:
+        return _cache[0]
+
+    from plugincode.location_provider import get_location
+
+    # try the environment first
+    magicdb_loc = os.environ.get(TYPECODE_LIBMAGIC_DB_PATH_ENVVAR)
+
+    if TRACE and magicdb_loc:
+        logger_debug('get_magicdb_location:', 'got environ magicdb location:', magicdb_loc)
+
+    # try a plugin-provided path second
+    if not magicdb_loc:
+        magicdb_loc = get_location(TYPECODE_LIBMAGIC_DB)
+
+        if TRACE and magicdb_loc:
+            logger_debug('get_magicdb_location:', 'got plugin magicdb location:', magicdb_loc)
+
+    # try the PATH
+    if not magicdb_loc:
+        db = 'magic.mgc'
+        magicdb_loc = command.find_in_path(db)
+
+        if magicdb_loc:
+            warnings.warn(
+                'magicdb found in the PATH. '
+                'Install instead a typecode-libmagic plugin for best support.\n'
+                f'OR set the {TYPECODE_LIBMAGIC_DB_PATH_ENVVAR} environment variable.'
+            )
+
+        if TRACE and magicdb_loc:
+            logger_debug('get_magicdb_location:', 'got path magicdb location:', magicdb_loc)
+
+    if not magicdb_loc:
+        warnings.warn(
+            'Libmagic magic database not found. '
+            'A default will be used if possible. '
+            'Install instead a typecode-libmagic plugin for best support.\n'
+            f'OR set the {TYPECODE_LIBMAGIC_DB_PATH_ENVVAR} environment variable.'
+        )
+        return
+
+    _cache.append(magicdb_loc)
+    return magicdb_loc
 
 
 if TRACE:
@@ -164,11 +329,12 @@ class Detector(object):
         self.flags = flags
         self.cookie = _magic_open(self.flags)
         if not magic_db_location:
-            magic_db_location = get_location(TYPECODE_LIBMAGIC_DATABASE)
+            # Caveat emptor: this may be empty in which case a default will be tried
+            magic_db_location = get_magicdb_location()
 
-        # Note: this location must always be bytes on Python2 and 3, all OSes
-        if isinstance(magic_db_location, str):
-            magic_db_location = fsencode(magic_db_location)
+        # Note: this location must always be FS-encoded bytes on all OSes
+        if magic_db_location and not isinstance(magic_db_location, bytes):
+            magic_db_location = os.fsencode(magic_db_location)
 
         _magic_load(self.cookie, magic_db_location)
 
@@ -190,7 +356,7 @@ class Detector(object):
             # location string may therefore be mangled and the file not accessible
             # anymore by libmagic in some cases.
             try:
-                uloc = fsencode(location)
+                uloc = os.fsencode(location)
                 return  _magic_file(self.cookie, uloc)
             except:
                 # if all fails, read the start of the file instead
@@ -257,3 +423,9 @@ _magic_load = libmagic.magic_load
 _magic_load.restype = ctypes.c_int
 _magic_load.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
 _magic_load.errcheck = check_error
+
+_magic_version = libmagic.magic_version
+_magic_version.restype = ctypes.c_int
+_magic_version.argtypes = []
+
+libmagic_version = _magic_version()